+++ JudoPayDotNetIntegrationTests/IntegrationTestsBase.cs
@@ -8,7 +8,6 @@ namespace JudoPayDotNetIntegrationTests
 {
     public abstract class IntegrationTestsBase
     {
-        protected JudoPayApi JudoPayApiIridium;
         protected JudoPayApi JudoPayApiBase;
         protected JudoPayApi JudoPayApiElevated;
         protected JudoPayApi JudoPayApiThreeDSecure2;
@@ -17,7 +16,7 @@ namespace JudoPayDotNetIntegrationTests
         protected IntegrationTestsBase() 
         {
             ServicePointManager.SecurityProtocol = SecurityProtocolType.Tls12;
-            JudoPayApiBase = JudoPayApiIridium = JudoPaymentsFactory.Create(Configuration.JudoEnvironment,
+            JudoPayApiBase = JudoPaymentsFactory.Create(Configuration.JudoEnvironment,
                 Configuration.Token, Configuration.Secret);
             JudoPayApiElevated = JudoPaymentsFactory.Create(Configuration.JudoEnvironment,
                 Configuration.ElevatedPrivilegesToken, Configuration.ElevatedPrivilegesSecret);
@@ -239,17 +238,12 @@ namespace JudoPayDotNetIntegrationTests
                     PostCode = "W31 4HS",
                     CountryCode = 826
                 },
-                ClientIpAddress = "127.0.0.1",
-                CompanyName = "Test",
                 Currency = "GBP",
                 ExpiryDate = DateTimeOffset.Now.AddMinutes(30),
                 JudoId = Configuration.Judoid,
-                PartnerServiceFee = 10,
                 CancelUrl = "https://www.test.com",
                 SuccessUrl = "https://www.test.com",
-                Reference = "42421",
-                Status = WebPaymentStatus.Open,
-                TransactionType = TransactionType.PAYMENT,
+                YourPaymentReference = Guid.NewGuid().ToString(),
                 YourConsumerReference = "4235325"
             };
         }
@@ -269,19 +263,14 @@ namespace JudoPayDotNetIntegrationTests
                     PostCode = "W31 4HS",
                     CountryCode = 826
                 },
-                ClientIpAddress = "127.0.0.1",
-                CompanyName = "Test",
                 Currency = "GBP",
                 ExpiryDate = DateTimeOffset.Now.AddMinutes(30),
                 JudoId = Configuration.Judoid,
-                PartnerServiceFee = 10,
-                CancelUrl = "https://www.test.com",
-                SuccessUrl = "https://www.test.com",
-                Reference = "42421",
-                Status = WebPaymentStatus.Open,
-                TransactionType = TransactionType.PAYMENT,
+                CancelUrl = "https://www.test.com/cancel",
+                SuccessUrl = "https://www.test.com/success",
+                YourPaymentReference = Guid.NewGuid().ToString(),
                 YourConsumerReference = "4235325",
-                MobileNumber = "07999999999",
+                MobileNumber = "7999999999",
                 PhoneCountryCode = "44",
                 EmailAddress = "test@judopay.com",
                 ThreeDSecure = new ThreeDSecureTwoModel()